Implemented MutationObserver over polling
