+++ docs/Changelog.md
@@ -2,7 +2,7 @@
 title: Roadmap and Changelog
 description: Changelog and feature roadmap for Quartz Syncer.
 created: 2025-05-16T12:59:31Z+0200
-modified: 2025-06-18T20:07:24Z+0200
+modified: 2025-06-18T22:20:09Z+0200
 publish: true
 ---
 
@@ -26,6 +26,7 @@ publish: true
 ### Version 1.8.2
 
 - Added support for the [Fantasy Statblocks](https://github.com/javalent/fantasy-statblocks) plugin.
+- Removed polling in favor of `MutationObserver`.
 - Updated documentation.
 
 ### Version 1.8.1
